agent-a/requirements.txt
anthropic>=0.25.0
apscheduler>=3.10.4
httpx>=0.27.0
python-dotenv>=1.0.0
